+++ CHANGELOG.md
@@ -9,7 +9,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Improve the output when running tasks.
 - Remove the destination folder if the copy process or one of the tasks fail.
 - Add a `cleanup_on_error` flag to optionally disable the cleanup feature.
-
+- Add the `skip_if_exists` option to skip files, without asking, if they already exists in the destination folder.
 
 ### Version 2.4.2 (2019-06)
 - Fix MAJOR bug that was preventing the `_exclude`, `_include` and `_tasks` keys from